+++ src/main/java/com/example/k_dish/model/repositories/DishRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DishRepository extends JpaRepository<Dish, Integer> {
+public interface DishRepository extends JpaRepository<Dish, Long> {
     List<Dish> findByName(String name);
     List<Dish> findByCost(float cost);
 }
