+++ src/main/java/com/example/k_dish/service/Service.java
@@ -4,8 +4,8 @@ import com.example.k_dish.model.entity.AbstractEntity;
 import java.util.List;
 
 public interface Service<T extends AbstractEntity> {
-    T read(Integer id);
+    T read(Long id);
     List<T> read();
     void save(T entity);
-    void delete(Integer id);
+    void delete(Long id);
 }
